% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county10}
\alias{county10}
\title{2010 Census gazetteer files for counties}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3221 rows and 12 columns.}
\source{
\url{https://www.census.gov/geo/maps-data/data/gazetteer2010.html}
}
\usage{
county10
}
\description{
From the Census website: "The counties file contains data for primary legal
divisions or statistically equivalent entities in the 50 states, District of
Columbia and Puerto Rico as of the 2010 Census."
}
\keyword{datasets}

