% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zcta10}
\alias{zcta10}
\title{2010 Census gazetteer files for zctas}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 33120 rows and 9 columns.}
\source{
\url{https://www.census.gov/geo/maps-data/data/gazetteer2010.html}
}
\usage{
zcta10
}
\description{
From the Census website: "The ZIP Code Tabulation Area file contains data for
all ZIP Code Tabulation Areas (ZCTAs) in the 50 states, the District of
Columbia and Puerto Rico as of the January 1, 2010."
}
\keyword{datasets}

